Remove |tee from start in late_command; add export to start.sh

+++ Virtual/Machine/ISO-maker/Files/For-Target/install-emacs.sh
@@ -9,7 +9,7 @@ sudo apt -y install emacs
 # done
 
 [[ -e /home/$myuser/.emacs ]] && sudo rm -f /home/$myuser/.emacs
-[[ -e          /root/.emacs ]] && sudo rm -f          /root/.emacs
+[[ -e          /root/.emacs ]]&& sudo rm -f          /root/.emacs
 
 cat /var/local/dotemacs_root_and_regular_users /var/local/dotemacs_regular_users_only > /var/local/dotemacs
 
@@ -17,8 +17,8 @@ sudo ln -s /var/local/dotemacs /home/$myuser/.emacs
 sudo ln -s /var/local/dotemacs_root_and_regular_users /root/.emacs
 
 # Make it clear in /var/local, where its content is used
-sudo ln -s /home/$myuser/.emacs /var/local/dotemacs-home 
-sudo ln -s /root/.emacs          /var/local/dotemacs-root
+[[ ! -e /var/local/dotemacs-home ]] && sudo ln -s /home/$myuser/.emacs /var/local/dotemacs-home
+[[ ! -e /var/local/dotemacs-root ]] && sudo ln -s /root/.emacs         /var/local/dotemacs-root
 
 # Permissions 
 chown "root:root" /root/.emacs                # no sudo
@@ -38,9 +38,9 @@ chmod a+rw /home/$myuser/.emacs.d
 
 echo 'keyserver hkp://keys.gnupg.net' > /home/$myuser/.emacs.d/elpa/gnupg/gpg.conf
 sudo -i -u  $myuser gpg --list-keys 
-sudo -i -u  $myuser gpg --homedir /home/$myuser/.emacs.d/elpa       --list-keys
+#sudo -i -u  $myuser gpg --homedir /home/$myuser/.emacs.d/elpa       --list-keys
 sudo -i -u  $myuser gpg --homedir /home/$myuser/.emacs.d/elpa/gnupg --list-keys
-sudo -i -u  $myuser gpg --homedir /home/$myuser/.emacs.d/elpa       --receive-keys 066DAFCB81E42C40
+#sudo -i -u  $myuser gpg --homedir /home/$myuser/.emacs.d/elpa       --receive-keys 066DAFCB81E42C40
 sudo -i -u  $myuser gpg --homedir /home/$myuser/.emacs.d/elpa/gnupg --receive-keys 066DAFCB81E42C40
 
 # Do emacs first-time setup (including downloading packages)
